Tidied up services code, all now using Heroku API, removed calcTS-mod.ts
