+++ src/containers/HelpCalc.jsx
@@ -1,4 +1,3 @@
-import styled from "@emotion/styled";
 import {
     checkBoxLabels,
     helpProcessor,
