+++ src/containers/SuperCalc.jsx
@@ -1,4 +1,3 @@
-import styled from "@emotion/styled";
 import Calculator from "./Calculator";
 import {
     inputLabels,
