+++ src/services/fireService.js
@@ -2,7 +2,7 @@ import { fetchFromApi } from "./utils";
 
 export const inputLabels = [
     "Age",
-    "Gross Income",
+    "Income",
     "Expenses",
     "Super Contribution Rate",
     "Deductions",
@@ -13,35 +13,17 @@ export const inputLabels = [
     "Assumed Growth",
 ];
 
-export const checkBoxLabels = ["Super Inclusive", "Private Hospital Cover"];
+export const checkBoxLabels = [
+    "Super Inclusive",
+    "Private Hospital Cover",
+    "Max Super Contributions",
+];
 
-export const resultLabels = ["Years To FIRE", "Age At FIRE"];
+export const resultLabels = [
+    "Years To FIRE",
+    // "Age At FIRE"
+];
 
 export const fireProcessor = async (formData) => {
-    // let formDataVals = {};
-    // for (const [key, value] of formData.entries()) {
-    //     formDataVals[key] =
-    //         Number(value) || value === "0" ? Number(value) : value;
-    // }
-    // console.log("Form Data:", formDataVals);
-
-    // const netWorth = new NetWorth(
-    //     formDataVals.grossIncome,
-    //     formDataVals.helpBalance,
-    //     formDataVals.superBalance,
-    //     formDataVals.investmentsBalance,
-    // );
-    // console.log(netWorth);
-    // const incomeTaxProfile = new IncomeTaxProfile(
-    //     netWorth,
-    //     formDataVals.expenses,
-    //     Boolean(formDataVals.superInclusive),
-    //     formDataVals.superContributionRate,
-    //     formDataVals.deductions,
-    //     formDataVals.fringeBenefits,
-    //     Boolean(formDataVals.privateHospitalCover),
-    // );
-    // console.log(incomeTaxProfile);
-    // return timeToFIRE(incomeTaxProfile, formDataVals.age, formDataVals.growth);
     return await fetchFromApi("fire", formData);
 };
