+++ src/services/helpService.js
@@ -11,10 +11,5 @@ export const checkBoxLabels = [];
 export const resultLabels = ["HELP Repayment Rate", "HELP Repayment"];
 
 export const helpProcessor = async (formData) => {
-    // const income = Number(formData.get("taxableIncome"));
-    // const rSuper = Number(formData.get("reportableSuperContributions"));
-    // const rFbt = Number(formData.get("reportableFringeBenefits"));
-    // console.log("Income: ", income, "FBT: ", rFbt, "Super: ", rSuper);
-    // return totalHELP(income + rFbt + rSuper);
     return await fetchFromApi("help-repay", formData);
 };
