+++ src/services/mlsService.js
@@ -7,13 +7,5 @@ export const checkBoxLabels = [];
 export const resultLabels = ["MLS Rate", "Medicare Levy Surcharge"];
 
 export const mlsProcessor = async (formData) => {
-    // const income = Number(formData.get("taxableIncome"));
-    // const rFbt = Number(formData.get("reportableFringeBenefits"));
-    // const body = {
-    //     income: income + rFbt,
-    // };
-    // const newForm = new FormData(body);
-    // console.log("Income: ", income, "FBT: ", rFbt);
-    // return totalMLS(income + rFbt);
     return await fetchFromApi("mls", formData);
 };
