+++ src/services/netIncomeService.js
@@ -1,21 +1,21 @@
-import { fetchFromApi } from "./utils";
+import { fetchFromApi, flattenResult } from "./utils";
 
 export const inputLabels = [
-    "Gross Income", //
-    "Expenses", //
-    "Super Contribution Rate", //
-    "Deductions", //
-    "Fringe Benefits", //
-    "HELP Balance", //
-    "Super Balance", //
-    "Investments Balance", //
-    "Assumed Growth", //
+    "Gross Income",
+    "Expenses",
+    "Super Contribution Rate",
+    "Deductions",
+    "Fringe Benefits",
+    "HELP Balance",
+    "Super Balance",
+    "Investments Balance",
+    "Assumed Growth",
 ];
 
 export const checkBoxLabels = [
-    "Super Inclusive", //
-    "Max Super Contributions", //
-    "Private Hospital Cover", //
+    "Super Inclusive",
+    "Max Super Contributions",
+    "Private Hospital Cover",
 ];
 
 export const resultLabels = [
@@ -37,48 +37,8 @@ export const resultLabels = [
 ];
 
 export const formProcessor = async (formData) => {
-    // let formDataVals = {};
-    // for (const [key, value] of formData.entries()) {
-    //     formDataVals[key] =
-    //         Number(value) || value === "0" ? Number(value) : value;
-    // }
-    // console.log("Form Data:", formDataVals);
-
-    // const netWorth = new NetWorth(
-    //     formDataVals.grossIncome,
-    //     formDataVals.helpBalance,
-    //     formDataVals.superBalance,
-    //     formDataVals.investmentsBalance,
-    // );
-    // console.log(netWorth);
-    // const incomeTaxProfile = new IncomeTaxProfile(
-    //     netWorth,
-    //     formDataVals.expenses,
-    //     Boolean(formDataVals.superInclusive),
-    //     formDataVals.superContributionRate,
-    //     formDataVals.deductions,
-    //     formDataVals.fringeBenefits,
-    //     Boolean(formDataVals.privateHospitalCover),
-    // );
-    // console.log(incomeTaxProfile);
-    // return taxTime(
-    //     incomeTaxProfile,
-    //     Boolean(formDataVals.max),
-    //     formDataVals.growth,
-    // );
+    // result has nested objects
     const nestedResult = await fetchFromApi("detailed-tax", formData);
-    const result = {};
 
-    // need to flatten nested object
-    for (const [key, value] of Object.entries(nestedResult)) {
-        if (typeof value != "object") {
-            result[key] = value;
-        } else {
-            for (const [subkey, subval] of Object.entries(value)) {
-                result[subkey] = subval;
-            }
-        }
-    }
-
-    return result;
+    return flattenResult(nestedResult);
 };
